vendor_backend/requirements.txt
fastapi
uvicorn
sqlite3
pydantic
python-slugify==8.0.1   